fashion-vision-ai/requirements.txt
# FastAPI & Server
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
python-dotenv==1.0.1
jinja2==3.1.5

# ML / Deep Learning
ultralytics==8.3.56
torch>=2.1.0
torchvision>=0.16.0
timm==1.0.12

# Image Processing
Pillow>=10.0.0
opencv-python-headless>=4.8.0
numpy>=1.24.0
scikit-learn>=1.3.0

# HTTP Client (OpenRouter)
httpx==0.28.1

# Utilities
aiofiles==24.1.0
